requirements.txt
easydict
h5py
nltk
numpy
pandas
Pillow
tqdm
termcolor
pyyaml
